Consolidate derived entities hooks
